+++ tests/test_filtering.py
@@ -3,9 +3,15 @@ import pytest
 from sqlalchemy import Column
 from sqlalchemy.sql.sqltypes import Boolean, Integer
 
-from datapipe.compute import Catalog, Pipeline, Table, run_pipeline
-from datapipe.core_steps import BatchGenerate, BatchTransform
-from datapipe.datatable import DataStore
+from datapipe import (
+    BatchGenerate,
+    BatchTransform,
+    Catalog,
+    DataStore,
+    Pipeline,
+    Table,
+    run_pipeline,
+)
 from datapipe.store.database import TableStoreDB
 
 from .util import assert_datatable_equal