+++ tests/test_transform_meta.py
@@ -4,10 +4,7 @@ import pytest
 from pytest_cases import parametrize
 from sqlalchemy import Column, Integer
 
-from datapipe.core_steps import BatchTransformStep
-from datapipe.metastore import MetaTable
-from datapipe.store.database import DBConn
-from datapipe.types import MetaSchema
+from datapipe import BatchTransformStep, DBConn, MetaSchema, MetaTable
 
 
 def make_mt(name, dbconn, schema_keys) -> MetaTable: