+++ tests/util.py
@@ -1,8 +1,6 @@
 import pandas as pd
 
-from datapipe.datatable import DataTable
-from datapipe.store.table_store import TableStore
-from datapipe.types import DataDF, data_to_index
+from datapipe import DataDF, DataTable, TableStore, data_to_index
 
 
 def assert_idx_equal(a, b):
@@ -12,22 +10,22 @@ def assert_idx_equal(a, b):
     assert a == b
 
 
-def assert_df_equal(a: pd.DataFrame, b: pd.DataFrame, index_cols=['id']) -> bool:
+def assert_df_equal(a: pd.DataFrame, b: pd.DataFrame, index_cols=["id"]) -> bool:
     a = a.set_index(index_cols)
     b = b.set_index(index_cols)
 
     assert_idx_equal(a.index, b.index)
 
-    eq_rows = (a.sort_index() == b.sort_index()).all(axis='columns')
+    eq_rows = (a.sort_index() == b.sort_index()).all(axis="columns")
 
     if eq_rows.all():
         return True
 
     else:
-        print('Difference')
-        print('A:')
+        print("Difference")
+        print("A:")
         print(a.loc[-eq_rows])
-        print('B:')
+        print("B:")
         print(b.loc[-eq_rows])
 
         raise AssertionError